feature: crear sección Comunidad
