+++ pages/comunidad/antonio-polo.md
@@ -7,3 +7,7 @@ twitter:
 linkedin: https://www.linkedin.com/in/antoniopolodealvarado
 github: https://github.com/polodealvarado
 ---
+
+## Colaboración
+- Ayuda en la organización del [Hackathon Somos NLP 2022](/blog/hackathon-2022)
+- Gestión del patrocinio de IBiDat | [Hackathon Somos NLP 2022](/blog/hackathon-2022)
