+++ pages/comunidad/frida-ruiz.md
@@ -7,3 +7,14 @@ twitter: https://twitter.com/FridaRuh
 linkedin: https://linkedin.com/in/fridaruh
 github: https://github.com/fridaruh
 ---
+
+## Colaboración
+- Gestión del patrocinio de AI The New Sexy | [Hackathon Somos NLP 2023](/hackathon)
+- Entrevista a María Grandury en su podcast "AI The New Sexy"
+
+<a href="https://open.spotify.com/episode/38Y2K2gAe8IujRZOMjv1BH?si=S8VO6ifPT4WYWLbSLr91tg"
+    target="_blank">
+    <img alt="AI The New Sexy" width="650" height="365"
+        src="https://somosnlp.github.io/assets/images/patrocinios/AiTheNewSexy.PNG" />
+</a>
+
