+++ pages/comunidad/omar-espejel.md
@@ -9,4 +9,5 @@ github: https://github.com/omarespejel
 ---
 
 ## Colaboración
-- Ayuda en la organización del [Hackathon Somos NLP 2022](/blog/hackathon-2022)
+
+- Gran ayuda en la organización del [Hackathon Somos NLP 2022](/blog/hackathon-2022)
