--- pages/core-team/antonio-polo.md
@@ -1,9 +1,0 @@
----
-title: Antonio Polo
-description: ML Engineer & Data Scientist @Nymiz
-cover: https://avatars.githubusercontent.com/u/30154911
-website: 
-twitter:
-linkedin: https://www.linkedin.com/in/antoniopolodealvarado
-github: https://github.com/polodealvarado
----
