--- pages/core-team/frida-ruiz.md
@@ -1,9 +1,0 @@
----
-title: Frida Ruíz
-description: Data Scientist
-cover: https://media.licdn.com/dms/image/D4E03AQFfObbA7_BnEg/profile-displayphoto-shrink_800_800/0/1668997843548?e=1678924800&v=beta&t=Ok4QiuJ_fIBJyOxNcO81-wIsHuJXrdP7w9c39mK0sII
-website: http://fridaruh.com/ 
-twitter: https://twitter.com/FridaRuh
-linkedin: https://linkedin.com/in/fridaruh
-github: https://github.com/fridaruh
----
