--- pages/core-team/ignacio-martinez.md
@@ -1,9 +1,0 @@
----
-title: Ignacio Martínez Soriano
-description: Data Scientist Research @ Hospital "Rafael Méndez"
-cover: https://pbs.twimg.com/profile_images/1498703209912750085/u712_rAt_400x400.jpg
-website: https://nachuss.github.io/
-twitter: https://twitter.com/coloratto67
-linkedin: https://www.linkedin.com/in/imsoriano
-github: https://github.com/NachusS
----
