--- pages/core-team/isaac-rodriguez.md
@@ -1,9 +1,0 @@
----
-title: Isaac Rodríguez Bribiesca
-description: NLP Engineer @Botco.ai
-cover: https://avatars.githubusercontent.com/u/9711043
-website: https://isaacrodgz.github.io
-twitter:
-linkedin: https://www.linkedin.com/in/isaacrodgz
-github: https://github.com/IsaacRodgz
----
