--- pages/core-team/lewis-tunstall.md
@@ -1,14 +1,0 @@
----
-title: Lewis Tunstall
-description: ML Engineer @HuggingFace 
-cover: https://somosnlp.github.io/assets/images/core_team_lewis.jpeg
-website: 
-twitter: https://twitter.com/_lewtun
-linkedin: https://www.linkedin.com/in/lewis-tunstall
-github: https://github.com/lewtun
----
-
-<!-- Contribuciones:
-- 2 x AMA en Hackathon 2022
-- host del evento de Graphex
-- 5 libros para el equipo ganador del Hackathon 2022 -->
