--- pages/core-team/manuel-romero.md
@@ -1,15 +1,0 @@
----
-title: Manuel Romero
-description: Senior NLP/NLG Engineer @Narrativa.AI
-cover: https://pbs.twimg.com/profile_images/1590735731839963136/Fz4GC7pZ_400x400.jpg
-website: https://hf.co/mrm8488
-twitter: https://twitter.com/mrm8488
-linkedin: https://www.linkedin.com/in/manuel-romero-cs/
-github: https://github.com/mrm8488
----
-
-Manuel Romero es ingeniero informático. Ha trabajado varios años como desarrollador backend. Actualmente trabaja como Senior ML Engineer en Narrativa. Es el principal contribuidor del Hub de ML más importante con casi 500 modelos de IA. Ha participado y participa en iniciativas como BigScience y BigCode donde entrenan modelos de lenguaje para hacerlos públicos. Manuel trabaja actualmente en Narrativa, startup especializada en la generación automática de contenido por medio de Inteligencia Artificial que desarrolla soluciones de generación de contenidos para medios de comunicación y diferentes empresas.
-
-<!-- Contribuciones:
-- Taller inaugural Hackathon 2022
-- Taller inaugural Hackathon 2023 -->
