--- pages/core-team/maria-grandury.md
@@ -1,11 +1,0 @@
----
-title: María Grandury
-description: ML Research Engineer @neurocat.AI | Fundadora @SomosNLP_
-cover: https://pbs.twimg.com/profile_images/1584913293470273537/6u-Q8SJP_400x400.jpg
-website: https://mariagrandury.com
-twitter: https://twitter.com/mariagrandury
-linkedin: https://www.linkedin.com/in/mariagrandury
-github: https://github.com/mariagrandury
----
-
-María Grandury es una ingeniera e investigadora de Machine Learning especializada en las áreas del PLN y la fiabilidad de la IA. Es matemática y física y actualmente trabaja en neurocat, una startup con sede en Berlín, donde desarrolla métodos de explicabilidad (XAI) y ataques adversarios para evaluar la seguridad de modelos de ML. Además, es la fundadora de Somos NLP, una comunidad internacional cuya misión es democratizar y acelerar el avance del NLP en español. Es Hugging Face Fellow y forma parte de varios proyectos de investigación de PLN internacionales, además de colaborar activamente con diversas iniciativas de mujeres en IA.
