--- pages/core-team/omar-sanseviero.md
@@ -1,9 +1,0 @@
----
-title: Omar Sanseviero
-description: ML Engineer @HuggingFace 
-cover: https://pbs.twimg.com/profile_images/1106315906165157889/0Hxb1ESL_400x400.png
-website: https://osanseviero.github.io/hackerllama/
-twitter: https://twitter.com/osanseviero
-linkedin: https://www.linkedin.com/in/omarsanseviero
-github: https://github.com/osanseviero
----
