--- pages/core-team/oscar-cumbicus.md
@@ -1,13 +1,0 @@
----
-title: Oscar Cumbicus
-description: Professor and Researcher
-cover: https://oskrmiguel.github.io/profile/images/user-3.jpg
-website: https://oskrmiguel.github.io/profile/
-twitter: https://twitter.com/OscarCumbicus
-linkedin: https://www.linkedin.com/in/oscar-cumbicus-47095443/
-github: https://github.com/oskrmiguel
----
-
-<!-- Contribuciones:
-- Patrocinio UNL Hackathon 2022
-- Colaboración con FliSol Loja -->
