--- pages/core-team/yhary-arias.md
@@ -1,15 +1,0 @@
----
-title: Yhary Arias
-description: Data Engineer | Master in AI @UPV
-cover: https://pbs.twimg.com/profile_images/1503384459005112325/vaAzE5yL_400x400.jpg
-website: 
-twitter: https://twitter.com/yharyarias5
-linkedin: https://www.linkedin.com/in/yharyarias/
-github: https://github.com/yharyarias
----
-
-## Colaboración
-
-- Diseño del cartel del 8M (WIP)
-- Diseño del banner de Twitter del Hackathon Somos NLP 2023
-- Traducción de 2 tutoriales de Hugging Face
