+++ pages/patrocinios/ai-the-new-sexy.md
@@ -10,4 +10,11 @@ website: https://linktr.ee/aithenewsexy
 Comunidad y podcast en español donde nos encargamos de hablar sobre el impacto de la Inteligencia Artificial en el día a día y sus aplicaciones.
 
 ## Colaboración
+
 - Comunidad patrocinadora de bronce del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+
+<div class="flex justify-center">
+    <img alt="AI The New Sexy" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/AiTheNewSexy.PNG" />
+</div>
