+++ pages/patrocinios/argilla.md
@@ -10,4 +10,10 @@ website: https://www.argilla.io/
 Argilla es una plataforma open-source de gestión y etiquetado de datos para proyectos NLP.  
 
 ## Colaboración
+
 - Empresa patrocinadora de oro del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+<div class="flex justify-center">
+    <img alt="Argilla" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/Argilla.svg" />
+</div>
