+++ pages/patrocinios/big-onion.md
@@ -10,4 +10,10 @@ website:
 Big Onion es un equipo de personas experimentadas en innovación con foco en negocio digital capaces de auditar, diseñar, ejecutar y formar equipos para proyectos tecnológicos.
 
 ## Colaboración
+
 - Empresa patrocinadora de bronce del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+<div class="flex justify-center">
+    <img alt="BigOnion" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/BigOnion.jpg" />
+</div>
