+++ pages/patrocinios/datagenero.md
@@ -10,4 +10,10 @@ website: https://www.datagenero.org
 Observatorio de datos con perspectiva de género.
 
 ## Colaboración
+
 - Organización patrocinadora de bronce del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+<div class="flex justify-center">
+    <img alt="DataGenero" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/DataGenero.png" />
+</div>
