+++ pages/patrocinios/gli-unam.md
@@ -12,3 +12,8 @@ Somos un grupo de investigación de la UNAM especializado en el Procesamiento de
 ## Colaboración
 
 - Patrocinador de bronce del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+<div class="flex justify-center">
+    <img alt="GIL_UNAM" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/GIL_UNAM.jpeg" />
+</div>
