+++ pages/patrocinios/hugging-face.md
@@ -9,5 +9,11 @@ website: https://huggingface.co
 Hugging Face es una plataforma colaborativa y open-source de Machine Learning en el que la comunidad ha compartido más de 150 mil modelos, 25 mil datasets y 30 mil aplicaciones de ML, incluyendo Stable DIffusion, Bloom y GPT-J. Hugging Face ha creado varias bibliotecas de ML como transformers, diffusers, gradio, datasets y optimum, las cuales permiten trabajar en todo el proceso de ML, desde el preprocesamiento de datos y el entrenamiento de modelos hasta poner modelos en producción y construir aplicaciones de estos.
 
 ## Colaboración
+
 - Patrocinador de oro del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
 - Patrocinador de oro del [Hackathon Somos NLP 2022: PLN en Español](/blog/hackathon-2022)
+
+<div class="flex justify-center">
+    <img alt="HuggingFace" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/HuggingFace.svg" />
+</div>
