+++ pages/patrocinios/iic.md
@@ -11,9 +11,14 @@ website: https://www.iic.uam.es
 
 - Patrocinador de plata del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
 
-## Personas de X que han colaborado con Somos NLP
+## Personas del IIC que han colaborado con Somos NLP
 
 - Keynote de Alejandro Vaca "Tecnologías clave de ChatGPT: Reinforcement Learning from Human Feedback" | Hackathon Somos NLP 2023
 - AMA con Alejandro Vaca | Hackathon Somos NLP 2023
 - BiomedIA | Proyectos de la comunidad, Hackathon Somos NLP 2022
 - RigoBERTa | Hackathon Somos NLP 2022
+
+<div class="flex justify-center">
+    <img alt="IIC" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/iic.bmp" />
+</div>
