+++ pages/patrocinios/mujeres-tech.md
@@ -10,4 +10,10 @@ website: www.mujerestech.com
 Mujeres Tech es una asociación sin ánimo de lucro que nace a principios de 2015 con el objetivo de dar visibilidad a las mujeres en el sector digital, fomentar las vocaciones científico-tecnológicas entre las más jóvenes, y servir de plataforma para establecer una red entre las mujeres profesionales del sector en España.
 
 ## Colaboración
-- Empresa patrocinadora de bronce del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+- Asociación patrocinadora de bronce del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+<div class="flex justify-center">
+    <img alt="MujeresTech" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/MujeresTech.png" />
+</div>
