+++ pages/patrocinios/narrativa.md
@@ -10,5 +10,11 @@ website: narrativa.com
 Narrativa crea contenido automático con tecnología NLG/NLP para sectores tan diversos como medios de comunicación, marketing o ciencias de la salud (farmacéuticas, CROs...).
 
 ## Colaboración
+
 - Empresa patrocinadora de plata del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
 - Empresa patrocinadora de plata del [Hackathon Somos NLP 2022: PLN en Español](/blog/hackathon-2022)
+
+<div class="flex justify-center">
+    <img alt="Narrativa" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/Narrativa.png" />
+</div>
