+++ pages/patrocinios/nlp-spain.md
@@ -10,4 +10,10 @@ website: nlpspain.blogspot.com
 NLP Spain es una comunidad para disfrutar divulgando y aprendiendo sobre Procesamiento del Lenguaje Natural.
 
 ## Colaboración
+
 - Comunidad patrocinadora de bronce del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+<div class="flex justify-center">
+    <img alt="NLPSpain" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/NLPSpain.jpg" />
+</div>
