+++ pages/patrocinios/omdena-mx.md
@@ -1,0 +1,21 @@
+---
+title: Omdena Mexico
+description:
+cover: https://somosnlp.github.io/assets/images/logo_omdena_mexico_perfil.png
+twitter: 
+linkedin: 
+website: https://omdena.com/
+---
+
+## Colaboración
+
+- Taller "Validador de Veracidad de Noticias" | Proyectos de la comunidad
+
+## Menciones en nuestro blog
+
+- Validador de Veracidad de Noticias (/blog/validador-veracidad-noticias)
+
+<div class="flex justify-center">
+    <img alt="Omdena Mexico" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/logo_omdena_mexico_perfil.png" />
+</div>
