+++ pages/patrocinios/saturdays-ai.md
@@ -8,5 +8,11 @@ website: https://saturdays.ai/
 ---
 
 ## Colaboración
+
 - Empresa patrocinadora de bronce del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
 - Empresa patrocinadora de plata del [Hackathon Somos NLP 2022: PLN en Español](/blog/hackathon-2022)
+
+<div class="flex justify-center">
+    <img alt="Saturdays AI" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/SaturdaysAI.png" />
+</div>
