+++ pages/patrocinios/software-themis.md
@@ -10,10 +10,18 @@ website: https://www.themis.es
 Software Themis es el primer corrector de lenguaje sexista.
 
 ## Colaboración
-- Patrocinador de bronce del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+- Empresa patrocinadora de bronce del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
 
 ## Personas de Software Themis que han colaborado con Somos NLP
+
 - Inés García Sacristán, ponente en la [Mesa Redonda 8M: PLN en la lucha contra el sexismo](https://www.youtube.com/watch?v=5fOiLWXQ78c&list=PLTA-KAy8nxaCIWrABV_JOJbriszTKIRXC)
 
 ## Menciones en nuestro blog
+
 - [Iniciativas 8M](https://somosnlp.org/blog/iniciativas-8m)
+
+<div class="flex justify-center">
+    <img alt="SoftwareThemis" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/SoftwareThemis.png" />
+</div>
