+++ pages/patrocinios/uned-nlp.md
@@ -10,4 +10,10 @@ website: https://sites.google.com/view/nlp-uned/home
 UNED Research Group in NLP & IR
 
 ## Colaboración
-- Patrocinador de plata del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+- Grupo patrocinador de plata del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
+
+<div class="flex justify-center">
+    <img alt="UNEDNLP" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/UNEDNLP.png" />
+</div>
