+++ pages/patrocinios/unl.md
@@ -10,5 +10,11 @@ website: https://unl.edu.ec
 Es misión de la Universidad Nacional de Loja: la formación académica y profesional, con sólidas bases científicas y técnicas, pertinencia social y valores; la generación y aplicación de conocimientos científicos, tecnológicos y técnicos, que aporten al desarrollo integral del entorno y al avance de la ciencia; el fortalecimiento del pensamiento, la promoción, desarrollo y difusión de los saberes y culturas; y, la prestación de servicios especializados.
 
 ## Colaboración
+
 - Universidad patrocinadora de plata del [Hackathon Somos NLP 2023: Los LLMs hablan Español](/hackathon)
 - Universidad patrocinadora de plata del [Hackathon Somos NLP 2022: PLN en Español](/blog/hackathon-2022)
+
+<div class="flex justify-center">
+    <img alt="UNL" width="650" height="365" 
+    src="https://somosnlp.github.io/assets/images/patrocinios/UNL.png" />
+</div>
